Implement AUTHOR_BIO "name" and "cover" fields.
In the theme readme, there's are "name" and "cover" fields in the
AUTHOR_BIO example, but these two fields weren't implemented yet.

Instead, the author page was displaying the author entry through
the title Jinja2 filter (which doesn't work so well for surnames like
mine, which contains multiple capital letters ;) ), and the cover
image was being used for the avatar image. This change uses the
"name" and "cover" fields for these, respectfully.

The cover image on the author page will also gracefully fall back
to the site default if an author cover isn't specified as well.